+++ Application/Usecases/Command/CreateSubjectCommand.cs
@@ -9,7 +9,7 @@ namespace Application.Usecases.Command
         public string SubjectName { get; set; }
         public string Description { get; set; }
         public bool IsActive { get; set; } = true;
-        public float MinAverageScoreToPass { get; set; } = 5.0f;
+        public double MinAverageScoreToPass { get; set; } = 5.0f;
     }
 }
