+++ Application/Usecases/Command/UpdateSubjectCommand.cs
@@ -8,6 +8,6 @@ namespace Application.Usecases.Command
         public string SubjectName { get; set; }
         public string Description { get; set; }
         public bool IsActive { get; set; }
-        public float MinAverageScoreToPass { get; set; }
+        public double MinAverageScoreToPass { get; set; }
     }
 }
